# backend/services/ai-chatbot/openapi.yaml
swagger: '2.0'
info:
  title: Grassroots Football Management API
  description: API Gateway for Grassroots Football Management microservices
  version: 1.0.0
host: gfm-1ykrsotxghm8m.apigateway.grassroots-football-management.cloud.goog
...
schemes:
  - https
securityDefinitions:
  firebase:
    type: oauth2
    flow: "implicit"
    authorizationUrl: ""
    x-google-issuer: "https://securetoken.google.com/grassroots-football-management"
    x-google-jwks_uri: "https://www.googleapis.com/service_accounts/v1/metadata/x509/securetoken@system.gserviceaccount.com"
    x-google-audiences: "grassroots-football-management"
security:
  - firebase: []
paths:
...
  /carpool/offer:
    post:
      x-google-backend:
        address: https://microservice-carpool-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-carpool-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Offer a ride for a specific match
      description: Allows a user to offer a carpool ride for a match.
      operationId: offerRide
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Ride details
          required: true
          schema:
            type: object
            required:
              - id
              - driverName
              - seats
              - location
              - pickup
              - time
              - matchId
              - matchDetails
            properties:
              id:
                type: string
                description: The unique ID of the ride
              driverName:
                type: string
                description: The name of the driver
              seats:
                type: integer
                description: Number of available seats
              location:
                type: string
                description: Departure location
              pickup:
                type: string
                enum: ["Yes", "No"]
                description: Whether pickup stops are available
              time:
                type: string
                format: time
                description: Departure time
              matchId:
                type: string
                description: The unique ID of the match associated with the ride
              matchDetails:
                type: string
                description: Match details (e.g., "Cork City vs Dublin United - 2024-02-20")
      responses:
        '201':
          description: Ride added successfully
          schema:
            type: object
            properties:
              message:
                type: string
              ride_id:
                type: string
        '400':
          description: Missing required fields
        '500':
          description: Internal server error
  /carpool/rides:
    get:
      x-google-backend:
        address: https://microservice-carpool-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-carpool-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Retrieve available rides for a specific team
      description: Retrieves a list of all carpool rides for the given club, age group, and division.
      operationId: getRides
      produces:
        - application/json
      parameters:
        - in: query
          name: clubName
          type: string
          description: Name of the club
          required: true
        - in: query
          name: ageGroup
          type: string
          description: Age group (e.g., U12, U14)
          required: true
        - in: query
          name: division
          type: string
          description: Division (e.g., Premier, First Division)
          required: true
      responses:
        '200':
          description: List of available rides for the specified team
          schema:
            type: array
            items:
              type: object
              properties:
                id:
                  type: string
                  description: Unique identifier for the ride
                driverName:
                  type: string
                  description: Name of the driver offering the ride
                seats:
                  type: integer
                  description: Number of available seats
                location:
                  type: string
                  description: Pickup location for the ride
                pickup:
                  type: string
                  enum: ["Yes", "No"]
                  description: Indicates whether pickup is required
                time:
                  type: string
                  format: time
                  description: Time of departure
                matchId:
                  type: string
                  description: ID of the match this ride is for
                matchDetails:
                  type: string
                  description: Brief details about the match
        '400':
          description: Missing required query parameters
          schema:
            type: object
            properties:
              error:
                type: string
                example: Missing required query parameters
        '500':
          description: Internal server error
          schema:
            type: object
            properties:
              error:
                type: string
                example: Internal server error
  /carpool/request:
    post:
      x-google-backend:
        address: https://microservice-carpool-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-carpool-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Confirm a carpool ride request
      description: Allows a user to confirm a ride for a specific match. The ride request is automatically confirmed, the available seats are reduced, and the user's name is added to the carpool's passenger list.
      operationId: requestRide
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Ride confirmation details
          required: true
          schema:
            type: object
            required:
              - userName
              - ride_id
            properties:
              userName:
                type: string
                description: The name of the user requesting the ride
              ride_id:
                type: string
                description: The unique ID of the ride
      responses:
        '200':
          description: Ride confirmed successfully
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Ride confirmed"
              ride_id:
                type: string
                example: "df2673b6-26ae-46b7-9dfd-630af4fdab5a"
              user:
                type: string
                example: "Jamie Roche"
        '400':
          description: Missing required fields or no available seats
          schema:
            type: object
            properties:
              error:
                type: string
                example: "No available seats"
        '404':
          description: Ride not found
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Ride not found"
        '500':
          description: Internal server error
  /carpool/cancel:
    post:
      x-google-backend:
        address: https://microservice-carpool-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-carpool-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Cancel a carpool ride
      description: Allows a driver to cancel a ride they offered.
      operationId: cancelRide
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: The ride ID to be canceled
          required: true
          schema:
            type: object
            properties:
              rideId:
                type: string
                description: The unique ID of the ride to be canceled
      responses:
        '200':
          description: Ride cancelled successfully
        '400':
          description: Missing ride ID in request
        '404':
          description: Ride not found
        '500':
          description: Internal server error
 

 # Payments Microservice Routes
...